requirements.txt
Flask
Flask-Mail
python-dotenv
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib